New query for search with updated relevance tests
