+++ src/api/Elastic.Documentation.Api.Core/Search/SearchUsecase.cs
@@ -2,7 +2,6 @@
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
 
-using System.Diagnostics;
 using Microsoft.Extensions.Logging;
 
 namespace Elastic.Documentation.Api.Core.Search;
